Create README.md
A wall page is being created similar to Facebook news feed where users will be able to post messages and see the messages posted by other users.

This app is integrated with a login app. After successfully logging in or registering, a user will be redirected to localhost:8000/wall, where the wall is displayed. Once we get the messages functionality working, it will allow users to post comments for any message. Comments will be in a separate database table from messages. It will allow the user to delete his/her own messages. It will also allow users to LIKE messages and comments.